+++ R/sim-progression.R
@@ -207,12 +207,12 @@ make.targeted.crosses <- function(first.parents, second.parents, map1=0L, map2=0
   if (is.numeric(first.parents) && !is.integer(first.parents)) {
     numeric.fp <- first.parents
     first.parents <- as.integer(first.parents)
-    if (!all.equal(numeric.fp,first.parents)) { stop("Parent 1 indexes must be integers.") }
+    if (!isTRUE(all(numeric.fp==first.parents))) { stop("Parent 1 indexes must be integers.") }
   }
   if (is.numeric(second.parents) && !is.integer(second.parents)) {
     numeric.sp <- second.parents
     second.parents <- as.integer(second.parents)
-    if (!all.equal(numeric.sp,second.parents)) { stop("Parent 2 indexes must be integers.") }
+    if (!isTRUE(all(numeric.sp==second.parents))) { stop("Parent 2 indexes must be integers.") }
   }
 	return(.Call(SXP_make_targeted_crosses, sim.data$p, first.parents, second.parents,
 				 map1, map2, give.names, name.prefix, offspring, track.pedigree, give.ids,